+++ services/accountmanager/wallet/service.go
@@ -68,6 +68,9 @@ func New(ctx context.Context, params ...Parameter) (*Service, error) {
 		log = log.Level(parameters.logLevel)
 	}
 
+	// Warn about lack of slashing protection
+	log.Warn().Msg("The wallet account manager does not provide built-in slashing protection.  Please use the dirk account manager for production systems.")
+
 	stores := make([]e2wtypes.Store, 0, len(parameters.locations))
 	if len(parameters.locations) == 0 {
 		// Use default location.